telebot/requirements.txt
python-telegram-bot
python-dotenv
grpcio-tools